KA-161 Rename mergeItemDataWithFlags to createViewModel
