+++ onboarding-task/src/utils/createViewModel.js
@@ -1,0 +1,8 @@
+import memoize from 'memoizee';
+
+export const createViewModel = (item, flags) => ({
+  ...item.toJS(),
+  ...flags.toJS(),
+});
+
+export const memoizedCreateViewModel = memoize(createViewModel);
