--- onboarding-task/src/utils/mergeItemDataWithFlags.js
@@ -1,8 +1,0 @@
-import memoize from 'memoizee';
-
-export const mergeItemDataWithFlags = (item, flags) => ({
-  ...item.toJS(),
-  ...flags.toJS(),
-});
-
-export const memoizedMergeItemDataWithFlags = memoize(mergeItemDataWithFlags);
